Removed redundant files. Moved class code from hilogame.py to new file named controller_class.py. Added all information to the README.md file.
