+++ README.md
@@ -1,3 +1,38 @@
-edited test
-# cse210-02
-codie 
+# Hilo
+In the hilo game, you guess if the next card will be higher or lower
+than the previous card. If you are correct, you earn 100 points. Wrong 
+and you lose 75. You start with 300 points. The game ends either when 
+you say so or if you run out of points.
+
+## Running the Game
+---
+You need to have Python 3.8.0 or newer installed and running. Open the root folder (cse210-02) in a terminal. To play the game, copy and paste the following command into the terminal:
+```
+python3 code\hilogame.py
+```
+If that fails, try:
+
+```
+py code\hilogame.py
+```
+
+## Project Structure
+---
+The project files and folders are organized as follows:
+```
+root                    (project root folder)
++--code                 (folder containing source code and classes)
+  +-- hilogame.py       (program entry point)
++-- README.md           (general info)
+```
+
+## Required Technologies
+---
+* Python 3.8.0
+
+## Authors
+---
+* Jake Rammell
+* Bradley Payne
+* Codie Snow
+* Josh Dalton
